controllers/inbody/requirements.txt
paddlepaddle
paddleocr
psycopg2-binary
python-dotenv
Pillow
numpy
opencv-python
openai